refactor(util/i18n): Move i18n code to i18n util.
