+++ lib/components/form/user-settings-i18n.js
@@ -1,27 +1,9 @@
 import coreUtils from '@opentripplanner/core-utils'
 
-import { getFormattedPlaces } from '../../util/i18n'
+import { getFormattedPlaces, renderCoordinates } from '../../util/i18n'
 
 const { hasTransit, toSentenceCase } = coreUtils.itinerary
 const { matchLatLon } = coreUtils.map
-
-// TODO: move to otp-ui/packages/core-utils/src/map.js (along coordsToString() and stringToCoords())
-
-/**
- * Reformats a {lat, lon} object to be internationalized.
- */
-export function renderCoordinates(intl, place) {
-  const MAX_FRAC_DIGITS = 5
-
-  return {
-    lat: intl.formatNumber(place.lat, {
-      maximumFractionDigits: MAX_FRAC_DIGITS
-    }),
-    lon: intl.formatNumber(place.lon, {
-      maximumFractionDigits: MAX_FRAC_DIGITS
-    })
-  }
-}
 
 // TODO: move to otp-ui/packages/core-utils/src/query.js
