+++ lib/util/i18n.js
@@ -335,3 +335,21 @@ export function getSupportedModes(config, intl) {
 
   return supportedModes
 }
+
+// TODO: move to otp-ui/packages/core-utils/src/map.js (along coordsToString() and stringToCoords())
+
+/**
+ * Reformats a {lat, lon} object to be internationalized.
+ */
+export function renderCoordinates(intl, place) {
+  const MAX_FRAC_DIGITS = 5
+
+  return {
+    lat: intl.formatNumber(place.lat, {
+      maximumFractionDigits: MAX_FRAC_DIGITS
+    }),
+    lon: intl.formatNumber(place.lon, {
+      maximumFractionDigits: MAX_FRAC_DIGITS
+    })
+  }
+}
